requirements.txt
amqp==2.5.1
anyjson==0.3.3
beautifulsoup4==4.8.1
billiard==3.6.1.0
celery==4.3.0
dj-database-url==0.5.0
Django==2.2.6
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.0.1
django-gravatar2==1.4.2
django-heroku==0.3.1
django-sendgrid-v5==0.8.0
djangorestframework==3.10.3
future==0.18.0
gunicorn==19.9.0
importlib-metadata==0.23
kombu==4.6.5
more-itertools==7.2.0
plotly==4.2.1
psycopg2==2.8.3
python-decouple==3.1
python-http-client==3.2.1
pytz==2019.3
redis==3.3.11
retrying==1.3.3
sendgrid==6.1.0
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
vine==1.3.0
whitenoise==4.1.4
zipp==0.6.0
requests==2.22.0